requirements.txt
Flask==2.0.1
Pillow==8.3.1
gunicorn==20.1.0 